Move test to SlidingWindowIndexTest.java
