+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/indexes/SlidingWindowWithPredicateTest.java
@@ -20,9 +20,11 @@
 
 package com.apple.foundationdb.record.provider.foundationdb.indexes;
 
+import com.apple.foundationdb.half.Half;
+import com.apple.foundationdb.linear.HalfRealVector;
+import com.apple.foundationdb.linear.Metric;
 import com.apple.foundationdb.record.RecordMetaData;
 import com.apple.foundationdb.record.RecordMetaDataBuilder;
-import com.apple.foundationdb.record.IndexBuildProto;
 import com.apple.foundationdb.record.TupleRange;
 import com.apple.foundationdb.record.metadata.Index;
 import com.apple.foundationdb.record.metadata.IndexComparison;
@@ -35,17 +37,11 @@ import com.apple.foundationdb.record.metadata.expressions.KeyWithValueExpression
 import com.apple.foundationdb.record.provider.foundationdb.FDBRecordContext;
 import com.apple.foundationdb.record.provider.foundationdb.FDBRecordStoreTestBase;
 import com.apple.foundationdb.record.provider.foundationdb.IndexMaintainer;
-import com.apple.foundationdb.record.provider.foundationdb.OnlineIndexer;
-import com.apple.foundationdb.record.provider.foundationdb.PendingWriteQueueIndexingFactory;
 import com.apple.foundationdb.record.provider.foundationdb.VectorIndexScanBounds;
 import com.apple.foundationdb.record.provider.foundationdb.VectorIndexScanOptions;
-import com.apple.foundationdb.record.provider.foundationdb.queue.PendingWritesQueue;
 import com.apple.foundationdb.record.query.expressions.Comparisons;
 import com.apple.foundationdb.record.slidingwindowvector.TestRecordsSlidingWindowVectorProto;
 import com.apple.foundationdb.record.slidingwindowvector.TestRecordsSlidingWindowVectorProto.SlidingWindowVectorRecord;
-import com.apple.foundationdb.half.Half;
-import com.apple.foundationdb.linear.HalfRealVector;
-import com.apple.foundationdb.linear.Metric;
 import com.apple.foundationdb.tuple.Tuple;
 import com.apple.test.Tags;
 import com.google.common.collect.ImmutableList;
@@ -54,7 +50,6 @@ import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 
 import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -62,9 +57,6 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * Tests for the interaction between sliding window semantics and base index predicates,
@@ -295,66 +287,5 @@ class SlidingWindowWithPredicateTest extends FDBRecordStoreTestBase {
             assertWindowContains(2, 4);
             commit(context);
         }
-    }
-
-    @Test
-    void writeOnlyWithQueueRoutesUpdatesToQueue() throws Exception {
-        try (FDBRecordContext context = openContext()) {
-            openStore(context, 2, Direction.DESC, 5);
-            recordStore.markIndexWriteOnlyWithQueue(INDEX_NAME).join();
-            assertTrue(recordStore.isIndexWriteOnlyWithQueue(INDEX_NAME));
-
-            rec(1, 100, 10);
-            rec(2, 200, 20);
-
-            // The records themselves are persisted.
-            assertNotNull(recordStore.loadRecord(Tuple.from(1L)));
-            assertNotNull(recordStore.loadRecord(Tuple.from(2L)));
-            commit(context);
-        }
-
-        // The updates went to the pending queue rather than the index: the queue holds exactly the two deferred writes.
-        try (FDBRecordContext context = openContext()) {
-            openStore(context, 2, Direction.DESC, 5);
-            final Index index = recordStore.getRecordMetaData().getIndex(INDEX_NAME);
-            final Long queueSize = queueSize(context, index);
-            assertEquals(2L, queueSize == null ? 0L : queueSize,
-                    "both deferred writes should be sitting in the pending queue");
-            commit(context);
-        }
-
-        // Drain the queue and finish the build via the online indexer.
-        try (FDBRecordContext context = openContext()) {
-            openStore(context, 2, Direction.DESC, 5);
-            final Index index = recordStore.getRecordMetaData().getIndex(INDEX_NAME);
-            try (OnlineIndexer indexer = OnlineIndexer.newBuilder()
-                    .setRecordStore(recordStore)
-                    .setIndex(index)
-                    .setIndexingPolicy(OnlineIndexer.IndexingPolicy.newBuilder()
-                            .setUsePendingWriteQueue(List.of(index))
-                            .build())
-                    .build()) {
-                indexer.buildIndex(true);
-            }
-            commit(context);
-        }
-
-        // After the drain the index is readable, valid (DESC window=2 keeps both passing records), and the queue has
-        // been erased.
-        try (FDBRecordContext context = openContext()) {
-            openStore(context, 2, Direction.DESC, 5);
-            final Index index = recordStore.getRecordMetaData().getIndex(INDEX_NAME);
-            assertTrue(recordStore.isIndexReadable(index));
-            assertWindowContains(1, 2);
-            assertNull(queueSize(context, index), "the queue data should have been erased once the index became readable");
-            commit(context);
-        }
-    }
-
-    @Nullable
-    private Long queueSize(@Nonnull FDBRecordContext context, @Nonnull Index index) {
-        final PendingWritesQueue<IndexBuildProto.PendingWritesQueueEntry> queue =
-                PendingWriteQueueIndexingFactory.getIndexingQueue(recordStore, index);
-        return queue.getQueueSizeNoConflict(context).join();
     }
 }
